# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Task Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset a user's password (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Security Groups
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
      required:
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    ResetPasswordRequest:
      type: object
      properties:
        newPassword:
          type: string
      required:
        - newPassword
...
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        surname:
          type: string
        email:
          type: string
        isAdmin:
          type: boolean
        isActive:
          type: boolean
        securityGroups:
          type: array
          items:
            $ref: "#/components/schemas/SecurityGroupSummary"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - surname
        - email
        - isAdmin
        - isActive
        - securityGroups
        - createdAt
    CreateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        surname:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        isAdmin:
          type: boolean
        isActive:
          type: boolean
        securityGroupIds:
          type: array
          items:
            type: integer
      required:
        - firstName
        - surname
        - email
        - password
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        surname:
          type: string
        email:
          type: string
          format: email
        isAdmin:
          type: boolean
        isActive:
          type: boolean
        securityGroupIds:
          type: array
          items:
            type: integer
    SecurityGroupSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name
...